security set up

+++ controllers/users.js
@@ -23,19 +23,21 @@ function usersCreate(req, res) {
 }
 
 function usersUpdate(req, res) {
- User.findByIdAndUpdate(req.params.id, req.body, { new: true },  (err, user) => {
+  if(req.user._id.toString() !== req.params.id) return res.status(401).json({ message: "Invalid action" });
+  User.findByIdAndUpdate(req.params.id, req.body, { new: true },  (err, user) => {
    if (err) return res.status(500).json({ message: "Something went wrong." });
    if (!user) return res.status(404).json({ message: "User not found." });
    return res.status(200).json(user);
- });
+  });
 }
 
 function usersDelete(req, res) {
- User.findByIdAndRemove(req.params.id, (err, user) => {
+  if(req.user._id.toString() !== req.params.id) return res.status(401).json({ message: "Invalid action" });
+  User.findByIdAndRemove(req.params.id, (err, user) => {
    if (err) return res.status(500).json({ message: "Something went wrong." });
    if (!user) return res.status(404).json({ message: "User not found." });
    return res.status(204).send();
- });
+  });
 }
